requirements.txt
pymongo
retrying
Mastodon.py